apps/api/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
litellm==1.61.15
sqlmodel==0.0.22
psycopg[binary]==3.2.12
alembic==1.13.2
pytest==8.3.3
PyJWT>=2.8.0
sentry-sdk==1.45.1
redis==5.2.0
starlette==0.49.1
httpx==0.28.1
